widened binary_find input range to u128

+++ src/lib.rs
@@ -74,12 +74,12 @@ where
 /// Otherwise binary_find returns Range of all the consecutive values PartiallyEqual to the sought item:&T.
 /// When item was not found, then the returned_range will be empty and 
 /// returned_range.start (and end) will give the sort position where the item can be inserted.
-pub fn binary_find<T,F>(range:Range<usize>,probe: F, item:&T )  
-    -> Result<Range<usize>,usize> where T:PartialOrd, F:Fn(usize)->T { 
+pub fn binary_find<T,F>(range:Range<u128>, probe: F, item:&T )  
+    -> Result<Range<u128>,u128> where T:PartialOrd, F:Fn(u128)->T { 
 
     // binary search lands possibly anywhere within several matching items
     // closure `last` finds the end of their range   
-    let last = |idx:usize| -> usize { 
+    let last = |idx:u128| -> u128 { 
         let mut lastidx = idx+1;
         for i in idx+1..range.end { // move end up
             if item == &probe(i) { lastidx += 1; } else { break; }; 
@@ -87,7 +87,7 @@ pub fn binary_find<T,F>(range:Range<usize>,probe: F, item:&T )
         lastidx
     };
     // closure `first` finds the start of the range of the matching items  
-    let first = |idx:usize| -> usize {
+    let first = |idx:u128| -> u128 {
         let mut firstidx = idx;
         for i in (range.start..idx).rev() { // move start down
             if item == &probe(i) { firstidx -= 1; } else { break; };
